blobnom-socket/requirements.txt
python-socketio
uvicorn[standard]
redis
pydantic
fastapi
pytz
protobuf
grpcio-tools